backend/requirements.txt
ortools>=9.8.3296
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
flask>=3.0.0
flask-cors>=4.0.0
chardet>=5.0.0